requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
kaleido>=0.2.1
pyyaml>=6.0
pytest>=7.0.0
psutil>=5.9.0
scipy>=1.10.0
